Rename graph to graph-core
