+++ src/vscode-bicep-ui/apps/visual-designer/src/features/graph-core/atoms/index.ts
