+++ src/vscode-bicep-ui/apps/visual-designer/src/features/graph-core/components/CanvasBackground.tsx
