+++ src/vscode-bicep-ui/apps/visual-designer/src/features/graph-core/hooks/index.ts
